feat: fix some bugs and add pageInfo for query

+++ aurora-modules/file-boot/src/main/java/xyz/xcye/file/service/FileService.java
@@ -167,7 +167,7 @@ public class FileService {
         PageData<FileVO> pageData = new PageData<>();
         pageData.setResult(files);
         pageData.setPages(1);
-        pageData.setTotal(files.size());
+        pageData.setTotal((long) files.size());
         return pageData;
     }
